+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 
 - global loglevel can preserved on flash
 - logging reads log-level from `etc/log.cfg.txt`
+- plugin url class for auto-discovery-configuration
+- skeleton for `dashboard` module
 -